requirements.txt
streamlit
geopandas
pandas
plotly.express
folium
streamlit_folium
os